Add UInfixT to TH types (fixes #10522)
Summary:
UInfixT is like UInfixE or UInfixP but for types. Template Haskell splices can
use it to punt fixity handling to GHC when constructing types.

UInfixT is converted in compiler/hsSyn/Convert to a right-biased tree of HsOpTy,
which is already rearranged in compiler/rename/RnTypes to match operator
fixities.

This patch consists of (1) adding UInfixT to the AST, (2) implementing the
conversion and updating relevant comments, (3) updating pretty-printing and
library support, and (4) adding tests.

Test Plan: validate

Reviewers: goldfire, austin, bgamari

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D1088

GHC Trac Issues: #10522