Generate .dyn_o files for .hsig (fixes #10660)
In -dynamic-too mode, .dyn_o files were not being generated for .hsig
files. This is because .hsig files skipped the pipeline, which normally
handles this.

Fix this by using the pipeline for .hsig files, as with .hs files,
instead of calling compileEmptyStub directly. This is simpler and
cleaneer than duplicating the -dynamic-too logic in the branch for
HsigFile.